requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==3.6.4.0
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pdfkit==1.0.0
prompt-toolkit==3.0.31
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.41
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
